installer: split update.rs into update, self_update, version
