+++ installer/src/lib.rs
@@ -14,6 +14,7 @@ pub mod robot_scan;
 pub mod run;
 pub mod run_context;
 pub mod say;
+pub mod self_update;
 pub mod setup;
 pub mod spawn;
 pub mod sudo;
@@ -21,4 +22,5 @@ pub mod systemd_service;
 pub mod uninstall;
 pub mod update;
 pub mod venv_forward;
+pub mod version;
 pub mod wizards;
