+++ installer/src/update.rs
@@ -4,7 +4,7 @@
 
 use std::path::{Path, PathBuf};
 
-use anyhow::{bail, Result};
+use anyhow::Result;
 
 use crate::cli::{HardwareSetupArgs, InstallMode, UpdateArgs};
 use crate::install_record::{self, Installed};
@@ -14,21 +14,17 @@ use crate::probe::{capture, Probes};
 use crate::run::{self, Report};
 use crate::run_context::Ctx;
 use crate::say;
+use crate::self_update::{rollback, self_update_or_note};
 use crate::setup::{dimos_venv, system_config, system_packages, verify};
+use crate::version::{normalize, parse_latest_tag, parse_ls_remote, RELEASES};
 use crate::wizards::nvidia::jetson;
 use crate::wizards::unitree::g1::{self, G1Observed};
 use crate::wizards::Robot;
 
-pub const RELEASES: &str = "https://github.com/dimensionalOS/dimos/releases";
 /// The same override `scripts/install.sh` honors, so a LAN test serves both from one directory.
 const INSTALLER_URL_ENV: &str = "DIMOS_INSTALLER_URL";
-/// A pre-release ranks below every marker, so a final release takes the highest rank.
-const FINAL: u8 = 3;
 const NOT_VERIFIED: &str = "verify not run in dry-run: a real `dimos update` runs the checks";
 
-const DOWNLOAD_TIMEOUT_S: u64 = 600;
-const SHA_TIMEOUT_S: u64 = 120;
-const SWAP_TIMEOUT_S: u64 = 30;
 const GIT_TIMEOUT_S: u64 = 600;
 /// A cold `uv sync` on an Orin NX builds wheels from source; the same budget `setup` gives it.
 const SYNC_TIMEOUT_S: u64 = 3600;
@@ -128,149 +124,6 @@ fn g1_record(installed: &Installed) -> Option<HardwareSetupArgs> {
         sdk_path: None,
         blueprint: None,
     })
-}
-
-/// `https://github.com/x/y/releases/tag/v0.0.15` -> `0.0.15`.
-pub fn parse_latest_tag(url: &str) -> Option<String> {
-    let (_, tag) = url.trim().rsplit_once("/tag/")?;
-    let version = normalize(tag);
-    (!version.is_empty()).then_some(version)
-}
-
-/// The 40-hex object id `git ls-remote` prints before the ref name.
-pub fn parse_ls_remote(text: &str) -> Option<String> {
-    let sha = text.split_whitespace().next()?;
-    (sha.len() == 40 && sha.chars().all(|c| c.is_ascii_hexdigit())).then(|| sha.to_string())
-}
-
-fn normalize(tag: &str) -> String {
-    tag.trim().trim_start_matches('v').to_string()
-}
-
-/// PEP 440-lite: the release numbers, then a rank that sorts `0.0.14b1` below `0.0.14`.
-pub fn version_key(v: &str) -> (Vec<u32>, (u8, u32)) {
-    let stripped = normalize(v);
-    let (release, pre) = split_pre(&stripped);
-    let numbers = release.split('.').map(|p| p.parse().unwrap_or(0)).collect();
-    (numbers, pre)
-}
-
-/// `rc` first: it carries a `c`, so a later `a`/`b` split would cut a release candidate in half.
-fn split_pre(v: &str) -> (&str, (u8, u32)) {
-    for (marker, rank) in [("rc", 2u8), ("a", 0), ("b", 1)] {
-        if let Some((release, n)) = v.split_once(marker) {
-            return (release, (rank, n.parse().unwrap_or(0)));
-        }
-    }
-    (v, (FINAL, 0))
-}
-
-pub fn newer(current: &str, candidate: &str) -> bool {
-    version_key(candidate) > version_key(current)
-}
-
-pub fn release_base(version: Option<&str>, override_url: Option<&str>) -> String {
-    match (override_url, version) {
-        (Some(url), _) => url.trim_end_matches('/').to_string(),
-        (None, Some(v)) => format!("{RELEASES}/download/v{}", normalize(v)),
-        (None, None) => format!("{RELEASES}/latest/download"),
-    }
-}
-
-pub fn artifact(target: &str) -> String {
-    format!("dimos-{target}")
-}
-
-/// The swap renames a sibling file over the binary, so it only works on the installed copy.
-pub fn require_installed_exe(home: &Path, current_exe: &Path) -> Result<PathBuf> {
-    let installed = install_record::installed_bin(home);
-    if current_exe != installed {
-        bail!(
-            "running {}, not the installed copy: `{} update` updates the binary",
-            current_exe.display(),
-            installed.display()
-        );
-    }
-    Ok(installed)
-}
-
-/// Downloads beside the binary (one filesystem, so the swap is an atomic rename) and swaps only
-/// after the sha256 and a `--version` run on the new file both pass.
-pub fn self_update_stage(base: &str, target: &str, bin: &Path, bak: &Path) -> Stage {
-    let (new, sums) = (bin.with_extension("new"), bin.with_extension("new.sha256"));
-    fetch_actions(base, &artifact(target), &new, &sums)
-        .into_iter()
-        .chain(swap_actions(bin, bak, new, sums))
-        .fold(Stage::new("self-update", false), Stage::push)
-}
-
-fn fetch_actions(base: &str, asset: &str, new: &Path, sums: &Path) -> Vec<Action> {
-    let (new_s, sums_s) = (text(new), text(sums));
-    let (bin_url, sums_url) = (format!("{base}/{asset}"), format!("{base}/{asset}.sha256"));
-    vec![
-        Action::run(
-            &["curl", "-fsSL", "--retry", "3", &bin_url, "-o", &new_s],
-            DOWNLOAD_TIMEOUT_S,
-        ),
-        Action::run(
-            &["curl", "-fsSL", "--retry", "3", &sums_url, "-o", &sums_s],
-            SHA_TIMEOUT_S,
-        ),
-        Action::VerifySha256 {
-            file: new.to_path_buf(),
-            sums_file: sums.to_path_buf(),
-        },
-        Action::run(&["chmod", "0755", &new_s], SWAP_TIMEOUT_S),
-        Action::run(&[&new_s, "--version"], SWAP_TIMEOUT_S),
-    ]
-}
-
-/// Copy, not rename: the binary is never absent, so a failed swap still leaves a working `dimos`.
-fn swap_actions(bin: &Path, bak: &Path, new: PathBuf, sums: PathBuf) -> Vec<Action> {
-    vec![
-        Action::Copy {
-            from: bin.to_path_buf(),
-            to: bak.to_path_buf(),
-            mode: 0o755,
-        },
-        Action::Rename {
-            from: new,
-            to: bin.to_path_buf(),
-        },
-        Action::Remove {
-            path: sums,
-            sudo: false,
-        },
-    ]
-}
-
-/// Skipped with a note rather than fatal: the rest of `update` is a repair and should still run.
-fn self_update_or_note(
-    args: &UpdateArgs,
-    obs: &Observed,
-    probes: &Probes,
-    home: &Path,
-) -> (Stage, Option<String>) {
-    let empty = Stage::new("self-update", false);
-    let bin = match require_installed_exe(home, &probes.current_exe) {
-        Ok(bin) => bin,
-        Err(e) => return (empty, Some(format!("{e:#}"))),
-    };
-    let target = match probes.platform.target() {
-        Ok(target) => target,
-        Err(e) => return (empty, Some(format!("{e:#}"))),
-    };
-    let Some(candidate) = obs.latest.as_deref() else {
-        return (empty, None);
-    };
-    if !args.force && !newer(DIMOS_VERSION, candidate) {
-        return (empty, None);
-    }
-    let base = release_base(Some(candidate), override_url().as_deref());
-    (
-        self_update_stage(&base, target, &bin, &install_record::backup_bin(home)),
-        None,
-    )
 }
 
 /// Library re-pins from PyPI, dev fast-forwards and re-syncs; empty when the probe says it is there.
@@ -444,25 +297,6 @@ fn record(installed: &Installed, obs: &Observed, report: &Report, home: &Path) -
     install_record::save(home, &out)
 }
 
-/// Puts the kept `.bak` back through the same executor, so `--dry-run --rollback` prints the swap.
-pub fn rollback(ctx: &mut Ctx, home: &Path) -> Result<i32> {
-    let (bin, bak) = (
-        install_record::installed_bin(home),
-        install_record::backup_bin(home),
-    );
-    if !bak.exists() {
-        bail!("nothing to roll back: {} does not exist", bak.display());
-    }
-    let steps = Plan {
-        command: "update --rollback".to_string(),
-        stages: vec![Stage::new("rollback", true).push(Action::Rename { from: bak, to: bin })],
-        notes: Vec::new(),
-    };
-    let report = run::run(&steps, ctx)?;
-    report.print(ctx);
-    Ok(report.exit_code())
-}
-
 /// `dimos update`, as a list of calls.
 pub fn run(
     args: &UpdateArgs,
@@ -497,7 +331,7 @@ fn rollback_hint(report: &Report, home: &Path) {
     }
 }
 
-fn override_url() -> Option<String> {
+pub(crate) fn override_url() -> Option<String> {
     std::env::var(INSTALLER_URL_ENV)
         .ok()
         .filter(|s| !s.is_empty())
@@ -648,100 +482,6 @@ mod tests {
 
     fn names(plan: &Plan) -> Vec<&str> {
         plan.stages.iter().map(|s| s.name).collect()
-    }
-
-    #[test]
-    fn artifact_is_dimos_dash_target() {
-        assert_eq!(artifact("aarch64-linux-musl"), "dimos-aarch64-linux-musl");
-    }
-
-    #[test]
-    fn release_base_prefers_the_url_override_then_the_pinned_tag_then_latest() {
-        assert_eq!(
-            release_base(Some("0.0.15"), Some("http://10.0.0.5:8000/")),
-            "http://10.0.0.5:8000"
-        );
-        assert_eq!(
-            release_base(Some("v0.0.15"), None),
-            format!("{RELEASES}/download/v0.0.15")
-        );
-        assert_eq!(
-            release_base(None, None),
-            format!("{RELEASES}/latest/download")
-        );
-    }
-
-    #[test]
-    fn version_key_sorts_a_prerelease_below_its_final_release() {
-        assert!(version_key("0.0.14b1") < version_key("0.0.14"));
-        assert!(version_key("0.0.14") < version_key("0.0.15"));
-        assert!(version_key("0.0.14a2") < version_key("0.0.14b1"));
-        assert!(version_key("0.0.14b1") < version_key("0.0.14rc1"));
-    }
-
-    #[test]
-    fn newer_is_false_for_the_same_version() {
-        assert!(!newer("0.0.14", "0.0.14"));
-        assert!(newer("0.0.14", "0.0.15"));
-        assert!(!newer("0.0.15", "0.0.14b1"));
-    }
-
-    #[test]
-    fn parse_latest_tag_reads_the_version_out_of_the_redirect_url() {
-        assert_eq!(
-            parse_latest_tag("https://github.com/dimensionalOS/dimos/releases/tag/v0.0.15\n"),
-            Some("0.0.15".to_string())
-        );
-        assert_eq!(parse_latest_tag("https://github.com/x/y/releases"), None);
-    }
-
-    #[test]
-    fn parse_ls_remote_takes_the_object_id_not_the_ref_name() {
-        let line = "9f8b2c1d0e4a6b7c8d9e0f1a2b3c4d5e6f708192\trefs/heads/main\n";
-        assert_eq!(
-            parse_ls_remote(line),
-            Some("9f8b2c1d0e4a6b7c8d9e0f1a2b3c4d5e6f708192".to_string())
-        );
-        assert_eq!(parse_ls_remote("fatal: could not read"), None);
-    }
-
-    #[test]
-    fn self_update_swaps_only_after_the_sha256_and_the_version_run_pass() {
-        let bin = PathBuf::from("/home/u/.local/bin/dimos");
-        let bak = PathBuf::from("/home/u/.local/bin/dimos.bak");
-        let stage = self_update_stage("http://host/d", "x86_64-linux-musl", &bin, &bak);
-        let verify_at = stage
-            .actions
-            .iter()
-            .position(|a| matches!(a, Action::VerifySha256 { .. }))
-            .expect("a sha256 check");
-        let rename_at = stage
-            .actions
-            .iter()
-            .position(|a| matches!(a, Action::Rename { .. }))
-            .expect("a swap");
-        assert!(verify_at < rename_at);
-        assert!(argv_of(&stage)
-            .iter()
-            .any(|argv| argv == &["/home/u/.local/bin/dimos.new", "--version"]));
-    }
-
-    #[test]
-    fn self_update_copies_the_old_binary_before_the_rename_so_it_is_never_absent() {
-        let bin = PathBuf::from("/home/u/.local/bin/dimos");
-        let bak = bin.with_extension("bak");
-        let stage = self_update_stage("http://host/d", "x86_64-linux-musl", &bin, &bak);
-        let copy_at = stage
-            .actions
-            .iter()
-            .position(|a| matches!(a, Action::Copy { to, .. } if *to == bak))
-            .expect("a backup copy");
-        let rename_at = stage
-            .actions
-            .iter()
-            .position(|a| matches!(a, Action::Rename { to, .. } if *to == bin))
-            .expect("a swap");
-        assert!(copy_at < rename_at);
     }
 
     #[test]
@@ -1005,13 +745,6 @@ mod tests {
             .expect("load")
             .expect("recorded");
         assert_eq!(after.dimos_version, DIMOS_VERSION);
-    }
-
-    #[test]
-    fn rollback_without_a_backup_names_the_missing_file() {
-        let home = TmpDir::new("update-rollback");
-        let err = rollback(&mut ctx(home.path(), true), home.path()).unwrap_err();
-        assert!(format!("{err:#}").contains("dimos.bak"));
     }
 
     #[test]
